refactor: chain configuration
