+++ packages/widget/src/domain/types/chains/ledger.ts
@@ -1,151 +1,11 @@
 import type { Currency, Families } from "@ledgerhq/wallet-api-client";
-import type { CosmosChainsAssets } from "@sk-widget/providers/cosmos/chains/types";
+import type { SupportedSKChains } from "@sk-widget/domain/types/chains";
 import {
   CosmosNetworks,
   EvmNetworks,
   MiscNetworks,
   SubstrateNetworks,
 } from "@stakekit/common";
-import type { Chain } from "@stakekit/rainbowkit";
-import type { Chain as ChainRainbowKit } from "@stakekit/rainbowkit";
-
-export const supportedCosmosChains = [
-  CosmosNetworks.Akash,
-  CosmosNetworks.Cosmos,
-  CosmosNetworks.Juno,
-  CosmosNetworks.Kava,
-  CosmosNetworks.Osmosis,
-  CosmosNetworks.Stargaze,
-  CosmosNetworks.Onomy,
-  CosmosNetworks.Persistence,
-  CosmosNetworks.Axelar,
-  CosmosNetworks.Quicksilver,
-  CosmosNetworks.Agoric,
-  CosmosNetworks.BandProtocol,
-  CosmosNetworks.Bitsong,
-  CosmosNetworks.Chihuahua,
-  CosmosNetworks.Comdex,
-  CosmosNetworks.Crescent,
-  CosmosNetworks.Cronos,
-  CosmosNetworks.Cudos,
-  CosmosNetworks.FetchAi,
-  CosmosNetworks.GravityBridge,
-  CosmosNetworks.IRISnet,
-  CosmosNetworks.KiNetwork,
-  CosmosNetworks.MarsProtocol,
-  CosmosNetworks.Regen,
-  CosmosNetworks.Secret,
-  CosmosNetworks.Sentinel,
-  CosmosNetworks.Sommelier,
-  CosmosNetworks.Teritori,
-  CosmosNetworks.Umee,
-  CosmosNetworks.Coreum,
-  CosmosNetworks.Desmos,
-  CosmosNetworks.Dydx,
-  CosmosNetworks.Injective,
-  CosmosNetworks.Sei,
-  CosmosNetworks.Mantra,
-] as const;
-export type SupportedCosmosChains = (typeof supportedCosmosChains)[number];
-const supportedCosmosChainsSet = new Set(supportedCosmosChains);
-export type CosmosChainsMap = {
-  [Key in SupportedCosmosChains]: {
-    type: "cosmos";
-    skChainName: Key;
-    wagmiChain: Chain;
-    chain: CosmosChainsAssets;
-  };
-};
-
-const supportedEVMChains = [
-  EvmNetworks.AvalancheC,
-  EvmNetworks.Arbitrum,
-  EvmNetworks.Binance,
-  EvmNetworks.Celo,
-  EvmNetworks.Ethereum,
-  EvmNetworks.EthereumGoerli,
-  EvmNetworks.Harmony,
-  EvmNetworks.Optimism,
-  EvmNetworks.Polygon,
-  EvmNetworks.Viction,
-  EvmNetworks.EthereumHolesky,
-  EvmNetworks.Base,
-  EvmNetworks.Linea,
-  EvmNetworks.Core,
-  EvmNetworks.Sonic,
-] as const;
-const supportedEVMChainsSet = new Set(supportedEVMChains);
-type SupportedEvmChain = (typeof supportedEVMChains)[number];
-export type EvmChainsMap = {
-  [Key in SupportedEvmChain]: {
-    type: "evm";
-    skChainName: Key;
-    wagmiChain: Chain;
-  };
-};
-
-const supportedMiscChains = [
-  MiscNetworks.Near,
-  MiscNetworks.Tezos,
-  MiscNetworks.Solana,
-  MiscNetworks.Tron,
-  MiscNetworks.Ton,
-] as const;
-const supportedMiscChainsSet = new Set(supportedMiscChains);
-type SupportedMiscChains = (typeof supportedMiscChains)[number];
-export type MiscChainsMap = {
-  [Key in SupportedMiscChains]: {
-    type: "misc";
-    skChainName: Key;
-    wagmiChain: Chain;
-  };
-};
-
-const supportedSubstrateChains = [SubstrateNetworks.Polkadot] as const;
-const supportedSubstrateChainsSet = new Set(supportedSubstrateChains);
-type SupportedSubstrateChains = (typeof supportedSubstrateChains)[number];
-export type SubstrateChainsMap = {
-  [Key in SupportedSubstrateChains]: {
-    type: "substrate";
-    skChainName: Key;
-    wagmiChain: Chain;
-  };
-};
-
-export const isEvmChain = (chain: string): chain is SupportedEvmChain => {
-  return supportedEVMChainsSet.has(chain as SupportedEvmChain);
-};
-
-export const isSolanaChain = (chain: string): chain is SupportedMiscChains => {
-  return chain === MiscNetworks.Solana;
-};
-
-export const isTonChain = (chain: string): chain is SupportedMiscChains => {
-  return chain === MiscNetworks.Ton;
-};
-
-export const isTronChain = (chain: string): chain is SupportedMiscChains => {
-  return chain === MiscNetworks.Tron;
-};
-
-export const isSupportedChain = (chain: string): chain is SupportedSKChains => {
-  return (
-    isEvmChain(chain) ||
-    supportedCosmosChainsSet.has(chain as SupportedCosmosChains) ||
-    supportedMiscChainsSet.has(chain as SupportedMiscChains) ||
-    supportedSubstrateChainsSet.has(chain as SupportedSubstrateChains)
-  );
-};
-
-export type SupportedSKChains =
-  | SupportedCosmosChains
-  | SupportedEvmChain
-  | SupportedMiscChains
-  | SupportedSubstrateChains;
-
-/**
- * LEDGER LIVE
- */
 
 export type SupportedLedgerLiveFamilies = Extract<
   Families,
@@ -351,8 +211,3 @@ export type SupportedLedgerFamiliesWithCurrency = Record<
     }
   >
 >;
-
-export type SupportedSKChainsType = {
-  chainName: SupportedSKChains;
-  chainIcon?: ChainRainbowKit["iconUrl"];
-};
