+++ packages/widget/src/domain/types/wallets/index.ts
@@ -1,8 +1,9 @@
 import type { TokenString } from "@sk-widget/domain/types";
+import type { SupportedSKChainIds } from "@sk-widget/domain/types/chains";
 import type { SKWallet } from "./generic-wallet";
 
 export type SKExternalProviders = {
-  currentChain?: number;
+  currentChain?: SupportedSKChainIds;
   currentAddress: string;
   initToken?: TokenString;
   supportedChainIds?: number[];
