+++ packages/widget/src/providers/cosmos/cosmos-connector.ts
@@ -6,6 +6,7 @@ import type {
   MainWalletBase,
 } from "@cosmos-kit/core";
 import type { WCClient } from "@cosmos-kit/walletconnect";
+import type { CosmosChainsMap } from "@sk-widget/domain/types/chains/cosmos";
 import type { Wallet } from "@stakekit/rainbowkit";
 import { SignDoc, TxRaw } from "cosmjs-types/cosmos/tx/v1beta1/tx";
 import EventEmitter from "eventemitter3";
@@ -14,7 +15,6 @@ import { BehaviorSubject } from "rxjs";
 import type { Address, Chain } from "viem";
 import type { CreateConnectorFn } from "wagmi";
 import { createConnector } from "wagmi";
-import type { CosmosChainsMap } from "../../domain/types/chains";
 import { getStorageItem, setStorageItem } from "../../services/local-storage";
 import { waitForMs } from "../../utils";
 import type { ExtraProps } from "./cosmos-connector-meta";
