+++ packages/widget/src/providers/ledger/ledger-connector.ts
@@ -4,6 +4,10 @@ import {
   WindowMessageTransport,
   deserializeTransaction,
 } from "@ledgerhq/wallet-api-client";
+import {
+  type SupportedLedgerLiveFamilies,
+  ledgerChainPriority,
+} from "@sk-widget/domain/types/chains/ledger";
 import type {
   Chain,
   WalletDetailsParams,
@@ -16,11 +20,7 @@ import type { CreateConnectorFn } from "wagmi";
 import { createConnector } from "wagmi";
 import { images } from "../../assets/images";
 import { skNormalizeChainId } from "../../domain";
-import {
-  type SupportedLedgerLiveFamilies,
-  type SupportedSKChains,
-  ledgerChainPriority,
-} from "../../domain/types/chains";
+import type { SupportedSKChains } from "../../domain/types/chains";
 import type { InitParams } from "../../domain/types/init-params";
 import { isLedgerDappBrowserProvider } from "../../utils";
 import { type ExtraProps, configMeta } from "./ledger-live-connector-meta";
