+++ packages/widget/src/providers/sk-wallet/utils.ts
@@ -1,13 +1,11 @@
+import type { CosmosChainsMap } from "@sk-widget/domain/types/chains/cosmos";
+import type { EvmChainsMap } from "@sk-widget/domain/types/chains/evm";
+import type { MiscChainsMap } from "@sk-widget/domain/types/chains/misc";
+import type { SubstrateChainsMap } from "@sk-widget/domain/types/chains/substrate";
 import type { Hex } from "viem";
 import { numberToHex } from "viem";
 import type { Chain } from "wagmi/chains";
 import type { SKWallet } from "../../domain/types";
-import type {
-  CosmosChainsMap,
-  EvmChainsMap,
-  MiscChainsMap,
-  SubstrateChainsMap,
-} from "../../domain/types/chains";
 import type { EVMTx } from "../../domain/types/wallets/generic-wallet";
 import { TxType } from "../../domain/types/wallets/generic-wallet";
 import type { DecodedEVMTransaction } from "./validation";
